Fix API endpoint through which company is created
Changes:
- created an extra directory for entities and an entity for company(cosidered a better practice to have entities and DTOs separated)
- added a table script for companies into already existing script(user_table.sql)
- corrected a few typographical errors
- reverted a few changes that could have affected the main branch

The endpoint returns the DTO of created company and status code of 201(Created)
The API's HTTP POST request was tested using Postman with the correct data

Note: unit tests are still missing, and the correctness of the data inside the code is not checked(e.g. if the object isn't null, etc). Also, the .sql script should be checked whether it is done by the requirements and the name should be reconsidered if so(currently user_table).